+++ lib/ui/page/home/tab/chats/widget/recent_chat.dart
@@ -40,7 +40,6 @@ import '/ui/page/home/widget/chat_tile.dart';
 import '/ui/page/home/widget/retry_image.dart';
 import '/ui/widget/context_menu/menu.dart';
 import '/ui/widget/svg/svg.dart';
-import '/ui/widget/widget_button.dart';
 import '/util/message_popup.dart';
 import '/util/platform_utils.dart';
 
@@ -138,17 +137,23 @@ class RecentChatTile extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    final ColorScheme colors = Theme.of(context).colorScheme;
+
     return Obx(() {
       final Chat chat = rxChat.chat.value;
       final bool isRoute = chat.isRoute(router.route, me);
+      final bool inverted = isRoute || selected;
 
       return ChatTile(
         chat: rxChat,
         status: [
-          _status(context),
+          _status(context, inverted),
           Text(
             chat.updatedAt.val.toLocal().toShort(),
-            style: Theme.of(context).textTheme.titleSmall,
+            style: Theme.of(context)
+                .textTheme
+                .titleSmall
+                ?.copyWith(color: inverted ? colors.onSecondary : null),
           ),
         ],
         subtitle: [
@@ -158,28 +163,32 @@ class RecentChatTile extends StatelessWidget {
             child: Row(
               children: [
                 const SizedBox(height: 3),
-                Expanded(child: _subtitle(context, selected)),
+                Expanded(child: _subtitle(context, selected, inverted)),
                 if (trailing == null) ...[
+                  _ongoingCall(context),
                   if (blocked) ...[
                     const SizedBox(width: 5),
-                    const Icon(
+                    Icon(
                       Icons.block,
-                      color: Color(0xFFC0C0C0),
+                      color: inverted
+                          ? colors.onSecondary
+                          : const Color(0xFFC0C0C0),
                       size: 20,
                     ),
-                    if (chat.muted == null) const SizedBox(width: 5),
-                  ],
-                  if (chat.muted != null) ...[
                     const SizedBox(width: 5),
-                    SvgImage.asset(
-                      'assets/icons/muted.svg',
+                  ] else if (chat.muted != null) ...[
+                    const SizedBox(width: 5),
+                    SvgLoader.asset(
+                      inverted
+                          ? 'assets/icons/muted_light.svg'
+                          : 'assets/icons/muted.svg',
                       key: Key('MuteIndicator_${chat.id}'),
                       width: 19.99,
                       height: 15,
                     ),
                     const SizedBox(width: 5),
                   ],
-                  _counter(),
+                  _counter(context, inverted),
                 ] else
                   ...trailing!,
               ],
@@ -236,7 +245,7 @@ class RecentChatTile extends StatelessWidget {
             trailing: const Icon(Icons.select_all),
           ),
         ],
-        selected: isRoute || selected,
+        selected: inverted,
         avatarBuilder: avatarBuilder,
         enableContextMenu: enableContextMenu,
         onTap: onTap ??
@@ -251,77 +260,11 @@ class RecentChatTile extends StatelessWidget {
 
   /// Builds a subtitle for the provided [RxChat] containing either its
   /// [Chat.lastItem] or an [AnimatedTyping] indicating an ongoing typing.
-  Widget _subtitle(BuildContext context, bool selected) {
+  Widget _subtitle(BuildContext context, bool selected, bool inverted) {
+    final ColorScheme colors = Theme.of(context).colorScheme;
+
     return Obx(() {
       final Chat chat = rxChat.chat.value;
-
-      if (chat.ongoingCall != null) {
-        final Widget trailing = WidgetButton(
-          key: inCall?.call() == true
-              ? const Key('DropCallButton')
-              : const Key('JoinCallButton'),
-          onPressed: inCall?.call() == true ? onDrop : onJoin,
-          child: Container(
-            padding: const EdgeInsets.all(4),
-            decoration: BoxDecoration(
-              borderRadius: BorderRadius.circular(20),
-              color: inCall?.call() == true
-                  ? Colors.red
-                  : Theme.of(context).colorScheme.secondary,
-            ),
-            child: LayoutBuilder(builder: (context, constraints) {
-              return Row(
-                mainAxisSize: MainAxisSize.min,
-                children: [
-                  const SizedBox(width: 8),
-                  Icon(
-                    inCall?.call() == true ? Icons.call_end : Icons.call,
-                    size: 16,
-                    color: Colors.white,
-                  ),
-                  const SizedBox(width: 8),
-                  if (constraints.maxWidth > 110)
-                    Expanded(
-                      child: Text(
-                        inCall?.call() == true
-                            ? 'btn_call_end'.l10n
-                            : 'btn_join_call'.l10n,
-                        style: Theme.of(context)
-                            .textTheme
-                            .titleSmall
-                            ?.copyWith(color: Colors.white),
-                        maxLines: 1,
-                        overflow: TextOverflow.ellipsis,
-                      ),
-                    ),
-                  const SizedBox(width: 8),
-                  PeriodicBuilder(
-                    period: const Duration(seconds: 1),
-                    builder: (_) {
-                      return Text(
-                        DateTime.now()
-                            .difference(chat.ongoingCall!.at.val)
-                            .hhMmSs(),
-                        style: Theme.of(context)
-                            .textTheme
-                            .titleSmall
-                            ?.copyWith(color: Colors.white),
-                      );
-                    },
-                  ),
-                  const SizedBox(width: 8),
-                ],
-              );
-            }),
-          ),
-        );
-
-        return DefaultTextStyle(
-          style: Theme.of(context).textTheme.titleSmall!,
-          overflow: TextOverflow.ellipsis,
-          child: AnimatedSwitcher(duration: 300.milliseconds, child: trailing),
-        );
-      }
 
       final ChatItem? item;
       if (rxChat.messages.isNotEmpty) {
@@ -359,7 +302,7 @@ class RecentChatTile extends StatelessWidget {
               draft.attachments.map((e) {
                 return Padding(
                   padding: const EdgeInsets.only(right: 2),
-                  child: _attachment(e),
+                  child: _attachment(e, inverted: inverted),
                 );
               }),
             );
@@ -367,7 +310,7 @@ class RecentChatTile extends StatelessWidget {
             images.add(
               Padding(
                 padding: const EdgeInsets.only(right: 4),
-                child: _attachment(draft.attachments.first),
+                child: _attachment(draft.attachments.first, inverted: inverted),
               ),
             );
           }
@@ -404,13 +347,13 @@ class RecentChatTile extends StatelessWidget {
                 Text(
                   'label_typing'.l10n,
                   style: TextStyle(
-                    color: Theme.of(context).colorScheme.secondary,
+                    color: inverted ? colors.onSecondary : colors.secondary,
                   ),
                 ),
                 const SizedBox(width: 3),
-                const Padding(
-                  padding: EdgeInsets.only(bottom: 4),
-                  child: AnimatedTyping(),
+                Padding(
+                  padding: const EdgeInsets.only(bottom: 4),
+                  child: AnimatedTyping(inverted: inverted),
                 ),
               ],
             ),
@@ -427,14 +370,14 @@ class RecentChatTile extends StatelessWidget {
                       maxLines: 1,
                       overflow: TextOverflow.ellipsis,
                       style: TextStyle(
-                        color: Theme.of(context).colorScheme.secondary,
+                        color: inverted ? Colors.white : colors.secondary,
                       ),
                     ),
                   ),
                   const SizedBox(width: 3),
-                  const Padding(
-                    padding: EdgeInsets.only(bottom: 4),
-                    child: AnimatedTyping(),
+                  Padding(
+                    padding: const EdgeInsets.only(bottom: 4),
+                    child: AnimatedTyping(inverted: inverted),
                   ),
                 ],
               ),
@@ -443,9 +386,13 @@ class RecentChatTile extends StatelessWidget {
         }
       } else if (item != null) {
         if (item is ChatCall) {
-          const Widget widget = Padding(
-            padding: EdgeInsets.fromLTRB(0, 2, 6, 2),
-            child: Icon(Icons.call, size: 16, color: Color(0xFF666666)),
+          Widget widget = Padding(
+            padding: const EdgeInsets.fromLTRB(0, 2, 6, 2),
+            child: Icon(
+              Icons.call,
+              size: 16,
+              color: inverted ? colors.onSecondary : const Color(0xFF666666),
+            ),
           );
 
           if (item.finishedAt == null && item.finishReason == null) {
@@ -476,6 +423,7 @@ class RecentChatTile extends StatelessWidget {
                     padding: const EdgeInsets.only(right: 2),
                     child: _attachment(
                       e,
+                      inverted: inverted,
                       onError: () => rxChat.updateAttachments(item!),
                     ),
                   );
@@ -487,6 +435,7 @@ class RecentChatTile extends StatelessWidget {
                   padding: const EdgeInsets.only(right: 4),
                   child: _attachment(
                     item.attachments.first,
+                    inverted: inverted,
                     onError: () => rxChat.updateAttachments(item!),
                   ),
                 ),
@@ -667,7 +616,10 @@ class RecentChatTile extends StatelessWidget {
       }
 
       return DefaultTextStyle(
-        style: Theme.of(context).textTheme.titleSmall!,
+        style: Theme.of(context)
+            .textTheme
+            .titleSmall!
+            .copyWith(color: inverted ? colors.onSecondary : null),
         overflow: TextOverflow.ellipsis,
         child: Row(children: subtitle),
       );
@@ -675,7 +627,11 @@ class RecentChatTile extends StatelessWidget {
   }
 
   /// Builds an [Attachment] visual representation.
-  Widget _attachment(Attachment e, {Future<void> Function()? onError}) {
+  Widget _attachment(
+    Attachment e, {
+    bool inverted = false,
+    Future<void> Function()? onError,
+  }) {
     Widget? content;
 
     if (e is LocalAttachment) {
@@ -695,19 +651,19 @@ class RecentChatTile extends StatelessWidget {
           );
         } else {
           content = Container(
-            color: Colors.grey,
-            child: const Icon(
+            color: inverted ? Colors.white : Colors.grey,
+            child: Icon(
               Icons.video_file,
               size: 18,
-              color: Colors.white,
+              color: inverted ? Colors.grey : Colors.white,
             ),
           );
         }
       } else {
         content = Container(
-          color: Colors.grey,
-          child: SvgImage.asset(
-            'assets/icons/file.svg',
+          color: inverted ? Colors.white : Colors.grey,
+          child: SvgLoader.asset(
+            inverted ? 'assets/icons/file_dark.svg' : 'assets/icons/file.svg',
             width: 30,
             height: 30,
           ),
@@ -742,19 +698,19 @@ class RecentChatTile extends StatelessWidget {
           );
         } else {
           content = Container(
-            color: Colors.grey,
-            child: const Icon(
+            color: inverted ? Colors.white : Colors.grey,
+            child: Icon(
               Icons.video_file,
               size: 18,
-              color: Colors.white,
+              color: inverted ? Colors.grey : Colors.white,
             ),
           );
         }
       } else {
         content = Container(
-          color: Colors.grey,
-          child: SvgImage.asset(
-            'assets/icons/file.svg',
+          color: inverted ? Colors.white : Colors.grey,
+          child: SvgLoader.asset(
+            inverted ? 'assets/icons/file_dark.svg' : 'assets/icons/file.svg',
             width: 30,
             height: 30,
           ),
@@ -777,7 +733,9 @@ class RecentChatTile extends StatelessWidget {
   }
 
   /// Builds a [ChatItem.status] visual representation.
-  Widget _status(BuildContext context) {
+  Widget _status(BuildContext context, bool inverted) {
+    final ColorScheme colors = Theme.of(context).colorScheme;
+
     return Obx(() {
       final Chat chat = rxChat.chat.value;
 
@@ -807,10 +765,14 @@ class RecentChatTile extends StatelessWidget {
                         ? Icons.error_outline
                         : Icons.done,
             color: isRead
-                ? Theme.of(context).colorScheme.secondary
+                ? inverted
+                    ? colors.onSecondary
+                    : colors.secondary
                 : isError
                     ? Colors.red
-                    : Theme.of(context).colorScheme.primary,
+                    : inverted
+                        ? colors.onSecondary
+                        : colors.primary,
             size: 16,
           ),
         );
@@ -821,9 +783,12 @@ class RecentChatTile extends StatelessWidget {
   }
 
   /// Returns a visual representation of the [Chat.unreadCount] counter.
-  Widget _counter() {
+  Widget _counter(BuildContext context, bool inverted) {
     return Obx(() {
+      final ColorScheme colors = Theme.of(context).colorScheme;
+
       final Chat chat = rxChat.chat.value;
+      final bool muted = chat.muted != null;
 
       if (rxChat.unreadCount.value > 0) {
         return Container(
@@ -833,7 +798,11 @@ class RecentChatTile extends StatelessWidget {
           height: 23,
           decoration: BoxDecoration(
             shape: BoxShape.circle,
-            color: chat.muted == null ? Colors.red : const Color(0xFFC0C0C0),
+            color: muted
+                ? inverted
+                    ? colors.onSecondary
+                    : const Color(0xFFC0C0C0)
+                : Colors.red,
           ),
           alignment: Alignment.center,
           child: Text(
@@ -841,8 +810,12 @@ class RecentChatTile extends StatelessWidget {
             rxChat.unreadCount.value > 99
                 ? '99${'plus'.l10n}'
                 : '${rxChat.unreadCount.value}',
-            style: const TextStyle(
-              color: Colors.white,
+            style: TextStyle(
+              color: muted
+                  ? inverted
+                      ? colors.primary
+                      : Colors.white
+                  : Colors.white,
               fontSize: 11,
               fontWeight: FontWeight.w600,
             ),
@@ -854,6 +827,80 @@ class RecentChatTile extends StatelessWidget {
       }
 
       return const SizedBox(key: Key('NoUnreadMessages'));
+    });
+  }
+
+  /// Returns a visual representation of the [Chat.ongoingCall], if any.
+  Widget _ongoingCall(BuildContext context) {
+    return Obx(() {
+      final Chat chat = rxChat.chat.value;
+
+      if (chat.ongoingCall == null) {
+        return const SizedBox();
+      }
+
+      // Returns a rounded rectangular button representing an [OngoingCall]
+      // associated action.
+      Widget button(bool active) {
+        return DecoratedBox(
+          key: active
+              ? const Key('JoinCallButton')
+              : const Key('DropCallButton'),
+          position: DecorationPosition.foreground,
+          decoration: BoxDecoration(
+            border: Border.all(color: Colors.white, width: 0.5),
+            borderRadius: BorderRadius.circular(20),
+          ),
+          child: Material(
+            elevation: 0,
+            type: MaterialType.button,
+            borderRadius: BorderRadius.circular(20),
+            color:
+                active ? Colors.red : Theme.of(context).colorScheme.secondary,
+            child: InkWell(
+              borderRadius: BorderRadius.circular(20),
+              onTap: active ? onDrop : onJoin,
+              child: Padding(
+                padding: const EdgeInsets.fromLTRB(10, 4, 10, 4),
+                child: Row(
+                  children: [
+                    Icon(
+                      active ? Icons.call_end : Icons.call,
+                      size: 16,
+                      color: Colors.white,
+                    ),
+                    const SizedBox(width: 6),
+                    PeriodicBuilder(
+                      period: const Duration(seconds: 1),
+                      builder: (_) {
+                        final Duration duration =
+                            DateTime.now().difference(chat.ongoingCall!.at.val);
+                        final String text = duration.hhMmSs();
+
+                        return Text(
+                          text,
+                          style: Theme.of(context)
+                              .textTheme
+                              .titleSmall
+                              ?.copyWith(color: Colors.white),
+                        ).fixedDigits();
+                      },
+                    )
+                  ],
+                ),
+              ),
+            ),
+          ),
+        );
+      }
+
+      return Padding(
+        padding: const EdgeInsets.only(left: 5),
+        child: AnimatedSwitcher(
+          duration: 300.milliseconds,
+          child: button(inCall?.call() == true),
+        ),
+      );
     });
   }
 
@@ -877,7 +924,13 @@ class RecentChatTile extends StatelessWidget {
   }
 
   /// Returns the [child].
-  static Widget _defaultAvatarBuilder(Widget child) => child;
+  ///
+  /// Uses [GestureDetector] with a dummy [GestureDetector.onLongPress] callback
+  /// for discarding long presses on its [child].
+  static Widget _defaultAvatarBuilder(Widget child) => GestureDetector(
+        onLongPress: () {},
+        child: child,
+      );
 }
 
 /// Extension adding conversion from [DateTime] to its short text relative to