+++ lib/ui/page/home/tab/contacts/view.dart
@@ -17,6 +17,7 @@
 
 import 'dart:ui';
 
+import 'package:animated_size_and_fade/animated_size_and_fade.dart';
 import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_staggered_animations/flutter_staggered_animations.dart';
@@ -53,6 +54,7 @@ class ContactsTabView extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final Style style = Theme.of(context).extension<Style>()!;
+    final ColorScheme colors = Theme.of(context).colorScheme;
 
     return GetBuilder(
       key: const Key('ContactsTab'),
@@ -67,10 +69,7 @@ class ContactsTabView extends StatelessWidget {
         return Scaffold(
           appBar: CustomAppBar(
             border: c.search.value != null || c.selecting.value
-                ? Border.all(
-                    color: Theme.of(context).colorScheme.secondary,
-                    width: 2,
-                  )
+                ? Border.all(color: colors.secondary, width: 2)
                 : null,
             title: Obx(() {
               final Widget child;
@@ -100,7 +99,33 @@ class ContactsTabView extends StatelessWidget {
               } else if (c.selecting.value) {
                 child = Text('label_select_contacts'.l10n);
               } else {
-                child = Text('label_contacts'.l10n);
+                final Widget synchronization;
+
+                if (c.fetching.value == null && c.status.value.isLoadingMore) {
+                  synchronization = Padding(
+                    padding: const EdgeInsets.only(top: 2),
+                    child: Center(
+                      child: Text(
+                        'label_synchronization'.l10n,
+                        style: TextStyle(fontSize: 13, color: colors.primary),
+                      ),
+                    ),
+                  );
+                } else {
+                  synchronization = const SizedBox.shrink();
+                }
+
+                child = Column(
+                  mainAxisSize: MainAxisSize.min,
+                  children: [
+                    Text('label_contacts'.l10n),
+                    AnimatedSizeAndFade(
+                      sizeDuration: const Duration(milliseconds: 300),
+                      fadeDuration: const Duration(milliseconds: 300),
+                      child: synchronization,
+                    ),
+                  ],
+                );
               }
 
               return AnimatedSwitcher(duration: 250.milliseconds, child: child);
@@ -172,7 +197,7 @@ class ContactsTabView extends StatelessWidget {
           ),
           extendBodyBehindAppBar: true,
           body: Obx(() {
-            if (!c.contactsReady.value) {
+            if (c.status.value.isLoading) {
               return const Center(child: CustomProgressIndicator());
             }
 
@@ -358,7 +383,15 @@ class ContactsTabView extends StatelessWidget {
                                           ),
                                           index: i,
                                           enabled: !c.selecting.value,
-                                          child: child,
+
+                                          // Use a dummy
+                                          // [GestureDetector.onLongPress]
+                                          // callback for discarding long
+                                          // presses on the [child].
+                                          child: GestureDetector(
+                                            onLongPress: () {},
+                                            child: child,
+                                          ),
                                         ),
                                       );
                                     },
@@ -454,6 +487,8 @@ class ContactsTabView extends StatelessWidget {
     ContactsTabController c, {
     Widget Function(Widget)? avatarBuilder,
   }) {
+    final ColorScheme colors = Theme.of(context).colorScheme;
+
     return Obx(() {
       bool favorite = c.favorites.contains(contact);
 
@@ -462,11 +497,13 @@ class ContactsTabView extends StatelessWidget {
               ?.startsWith('${Routes.user}/${contact.user.value?.id}') ==
           true;
 
+      final bool inverted = selected || c.selectedContacts.contains(contact.id);
+
       return ContactTile(
         key: Key('Contact_${contact.id}'),
         contact: contact,
         folded: favorite,
-        selected: selected || c.selectedContacts.contains(contact.id),
+        selected: inverted,
         enableContextMenu: !c.selecting.value,
         avatarBuilder: c.selecting.value
             ? (child) => WidgetButton(
@@ -517,8 +554,9 @@ class ContactsTabView extends StatelessWidget {
               if (subtitle != null) {
                 return Text(
                   subtitle,
-                  style:
-                      TextStyle(color: Theme.of(context).colorScheme.primary),
+                  style: TextStyle(
+                    color: inverted ? colors.onSecondary : colors.primary,
+                  ),
                 );
               }
 
@@ -530,14 +568,17 @@ class ContactsTabView extends StatelessWidget {
           Obx(() {
             final dialog = contact.user.value?.dialog.value;
 
-            if (dialog?.chat.value.muted == null) {
+            if (dialog?.chat.value.muted == null ||
+                contact.user.value?.user.value.isBlacklisted != null) {
               return const SizedBox();
             }
 
             return Padding(
               padding: const EdgeInsets.symmetric(horizontal: 5),
-              child: SvgImage.asset(
-                'assets/icons/muted.svg',
+              child: SvgLoader.asset(
+                inverted
+                    ? 'assets/icons/muted_light.svg'
+                    : 'assets/icons/muted.svg',
                 key: Key('MuteIndicator_${contact.id}'),
                 width: 19.99,
                 height: 15,
@@ -549,9 +590,13 @@ class ContactsTabView extends StatelessWidget {
               return const SizedBox();
             }
 
-            return const Padding(
-              padding: EdgeInsets.symmetric(horizontal: 5),
-              child: Icon(Icons.block, color: Color(0xFFC0C0C0), size: 20),
+            return Padding(
+              padding: const EdgeInsets.symmetric(horizontal: 5),
+              child: Icon(
+                Icons.block,
+                color: inverted ? colors.onSecondary : const Color(0xFFC0C0C0),
+                size: 20,
+              ),
             );
           }),
           Obx(() {